requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.8.0
sentence-transformers>=2.0.0
faiss-cpu>=1.0.0
streamlit-audiorecorder>=0.0.2
